# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IoT Trap Monitoring System API",
        "description": "API для системы мониторинга IoT ловушек для борьбы с вредителями. Включает управление шлюзами, ловушками и регистрацию событий срабатывания.",
        "contact": {
            "email": "support@trap-monitoring.local"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:3000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.trap-monitoring.local/api",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/v1/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "List all customers",
                "operationId": "f2066eaaf1fa79d2e7658ef86e3566e4",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "individual",
                                "company"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of customers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CustomerResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "operationId": "141556723be774f5dcc8402358875c13",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCustomerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/v1/customers/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get customer by ID with facilities",
                "operationId": "cfeeafe51258aac882f26f8462e73df4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer details with facilities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerWithFacilitiesResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update customer",
                "operationId": "21e4130bfdfa4908f6d67921e8b9be61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCustomerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomerResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete customer (soft delete with cascade)",
                "description": "Soft deletes the customer, all its facilities, and sets all gateways and traps to IN_STOCK status",
                "operationId": "7fe6c4ed57f4ab6ca572b5f48bc9458a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Customer deleted successfully"
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                }
            }
        },
        "/v1/customers/{id}/facilities": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "List all facilities for a customer",
                "operationId": "a7de42d3077c75c94daefd18e60a3bfe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of facilities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FacilityResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateCustomerRequest": {
                "required": [
                    "name",
                    "email",
                    "phone",
                    "type",
                    "billing_address"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "ACME Corporation"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "info@acme.com"
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "+380501234567"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "individual",
                            "company"
                        ],
                        "example": "company"
                    },
                    "billing_address": {
                        "required": [
                            "street",
                            "city",
                            "postal_code",
                            "country"
                        ],
                        "properties": {
                            "street": {
                                "type": "string",
                                "example": "123 Business Ave"
                            },
                            "city": {
                                "type": "string",
                                "example": "Kyiv"
                            },
                            "postal_code": {
                                "type": "string",
                                "example": "01001"
                            },
                            "country": {
                                "type": "string",
                                "example": "Ukraine"
                            }
                        },
                        "type": "object"
                    },
                    "tax_number": {
                        "type": "string",
                        "example": "1234567890",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "example": "Premium customer",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateCustomerRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "ACME Corporation"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "info@acme.com"
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "+380501234567"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "individual",
                            "company"
                        ],
                        "example": "company"
                    },
                    "billing_address": {
                        "required": [
                            "street",
                            "city",
                            "postal_code",
                            "country"
                        ],
                        "properties": {
                            "street": {
                                "type": "string",
                                "example": "123 Business Ave"
                            },
                            "city": {
                                "type": "string",
                                "example": "Kyiv"
                            },
                            "postal_code": {
                                "type": "string",
                                "example": "01001"
                            },
                            "country": {
                                "type": "string",
                                "example": "Ukraine"
                            }
                        },
                        "type": "object"
                    },
                    "tax_number": {
                        "type": "string",
                        "example": "1234567890",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "example": "Premium customer",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    }
                },
                "type": "object"
            },
...
            "CustomerResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "billing_address": {
                        "properties": {
                            "street": {
                                "type": "string"
                            },
                            "city": {
                                "type": "string"
                            },
                            "postal_code": {
                                "type": "string"
                            },
                            "country": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "tax_number": {
                        "type": "string",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CustomerWithFacilitiesResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "billing_address": {
                        "properties": {
                            "street": {
                                "type": "string"
                            },
                            "city": {
                                "type": "string"
                            },
                            "postal_code": {
                                "type": "string"
                            },
                            "country": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "tax_number": {
                        "type": "string",
                        "nullable": true
                    },
                    "notes": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "facilities": {
                        "description": "Объекты, принадлежащие клиенту",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FacilityResource"
                        }
                    },
                    "facilities_count": {
                        "description": "Количество объектов",
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
            "FacilityResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "properties": {
                            "street": {
                                "type": "string"
                            },
                            "city": {
                                "type": "string"
                            },
                            "postal_code": {
                                "type": "string"
                            },
                            "country": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "contact_person": {
                        "type": "string"
                    },
                    "contact_phone": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "area": {
                        "type": "number",
                        "format": "float"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "HmacSignature": {
                "type": "apiKey",
                "description": "HMAC SHA-256 signature для защиты IoT endpoints. Используйте hash_hmac(\"sha256\", payload, gateway_api_key)",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
...
}